devices/bus/c64/exp: reworked cart loading [hap]

+++ src/devices/bus/c64/exp.h
@@ -153,8 +153,8 @@ protected:
 	std::unique_ptr<uint8_t[]> m_romx;
 	std::unique_ptr<uint8_t[]> m_nvram;
 
-	uint32_t m_roml_size;
-	uint32_t m_romh_size;
+	size_t m_roml_size;
+	size_t m_romh_size;
 
 	int m_game;
 	int m_exrom;
